pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gtts>=2.5.4",
    "gunicorn>=23.0.0",
    "markdown>=3.8",
    "moviepy>=2.2.1",
    "opencv-python>=4.11.0.86",
    "psycopg2-binary>=2.9.10",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "python-markdown>=0.1.0",
    "pyttsx3>=2.98",
    "pytube>=15.0.0",
    "reportlab>=4.4.1",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.40",
    "trafilatura>=2.0.0",
    "werkzeug>=3.1.3",
    "yt-dlp>=2025.6.9",
]
